# Implements serializables based on new serialization mechanism