--- a/mars/serialization/serializables/field_type.py
+++ b/mars/serialization/serializables/field_type.py
@@ -437,6 +437,8 @@ class DictType(AbstractFieldType):
 
     def validate(self, value):
         super().validate(value)
+        if value is None:
+            return
         for k, v in value.items():
             try:
                 self.key_type.validate(k)